feat(data): Add template data for entities
Using the Legion ECS system, we can load up templates of entity
data from disk. Using this data will prevent us from needing to
manually spawn most of the entities in the game. The exceptions
are the player and the Amulet of Yala, for which we only want one
entity.